weight + stats FINALLY COMPLETE!

+++ src/main/java/com/luna/synthesis/features/utilities/FindSomeonesSkyblockInfo.java
@@ -221,7 +221,7 @@ public class FindSomeonesSkyblockInfo {
                 String essenceString = (EnumChatFormatting.DARK_PURPLE + "Total essence: §r" + totalEssence + " Essence (of which they have " + EnumChatFormatting.BLUE + iceEssence + " Ice, " + EnumChatFormatting.GRAY + witherEssence + " Wither, " + EnumChatFormatting.DARK_RED + spiderEssence + " Spider, " + EnumChatFormatting.DARK_PURPLE + undeadEssence + " Undead, " + EnumChatFormatting.AQUA + diamondEssence + " Diamond, " + EnumChatFormatting.YELLOW + dragonEssence + " Dragon, " + EnumChatFormatting.GOLD + goldEssence + " Gold, and " + EnumChatFormatting.RED + crimsonEssence + " Crimson§r)");
                 double overallSenitherWeight = (((JsonObject)(currentSbProfileWeightData).getAsJsonObject().get("senither")).get("overall").getAsDouble());
                 double overallLilyWeight = (((JsonObject)(currentSbProfileWeightData).getAsJsonObject().get("lily")).get("total").getAsDouble());
-                String weightString = (EnumChatFormatting.YELLOW + "Overall Lily Weight: " + overallLilyWeight + linePrefix + "§eOverall Senither Weight: " + overallSenitherWeight);
+                String weightString = (EnumChatFormatting.YELLOW + "Overall Lily Weight: " + ((int)(overallLilyWeight)) + linePrefix + "§eOverall Senither Weight: " + ((int)(overallSenitherWeight)));
                 if (rankPrefix.equals("")) {
                     rankPrefix = ("§7");
                 } else {
